Move permissions to own module to avoid circular imports
Revert type checker change
